create-flask-app/api/requirements.txt
Flask
flask-cors
SQLAlchemy
psycopg[binary]
python-dotenv
requests
